Update The Good, The Bad, and The Ugly: Tales of an Unkept Web.md
I enjoyed it! I found a couple of small typos, which I corrected in the edit. I also added a couple of links to stuff I thought might be fun. But yeah, I'd say submit it. I'm sure you'll get more feedback from the crew but I think it's ready for that step. Nice job!